+++ Application/Exceptions/ValidationException.cs
@@ -1,0 +1,34 @@
+﻿using FluentValidation.Results;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+namespace Application.Exceptions
+{
+    public class ValidationException : Exception
+    {
+        public ValidationException()
+            : base("One or more validation failures have occurred.")
+        {
+            Errors = new Dictionary<string, string[]>();
+        }
+
+        public ValidationException(IEnumerable<ValidationFailure> failures)
+            : this()
+        {
+            var failureGroups = failures
+                .GroupBy(e => e.PropertyName, e => e.ErrorMessage);
+
+            foreach (var failureGroup in failureGroups)
+            {
+                var propertyName = failureGroup.Key;
+                var propertyFailures = failureGroup.ToArray();
+
+                Errors.Add(propertyName, propertyFailures);
+            }
+        }
+
+        public IDictionary<string, string[]> Errors { get; }
+    }
+}