+++ Application/Features/Products/Commands/CreateProduct/CreateProductCommandValidator.cs
@@ -1,0 +1,44 @@
+﻿using Application.Interfaces.Repositories;
+using Domain.Entities;
+using FluentValidation;
+using Microsoft.EntityFrameworkCore.Internal;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace Application.Features.Products.Commands.CreateProduct
+{
+    public class CreateProductCommandValidator : AbstractValidator<CreateProductCommand>
+    {
+        private readonly IProductRepositoryAsync productRepository;
+
+        public CreateProductCommandValidator(IProductRepositoryAsync productRepository)
+        {
+            this.productRepository = productRepository;
+
+            RuleFor(p => p.Barcode)
+                .NotEmpty().WithMessage("{PropertyName} is required.")
+                .NotNull()
+                .MaximumLength(50).WithMessage("{PropertyName} must not exceed 50 characters.")
+                .MustAsync(IsUniqueBarcode).WithMessage("{PropertyName} already exists.");
+
+            RuleFor(p => p.Name)
+                .NotEmpty().WithMessage("{PropertyName} is required.")
+                .NotNull()
+                .MaximumLength(50).WithMessage("{PropertyName} must not exceed 50 characters.");
+                
+        }
+        public async Task<bool> IsUniqueBarcode(string barcode, CancellationToken cancellationToken)
+        {
+            // TODO Check performance here
+            var products =  await productRepository.GetAllAsync();
+            if (products.Count == 0) return true;
+            var unique = products.Where(a => a.Barcode == barcode).ToList();
+            if (unique.Count > 0) return false;
+            return true;
+        }
+    }
+}