+++ Application/ServiceRegistration.cs
@@ -1,4 +1,6 @@
-﻿using AutoMapper;
+﻿using Application.Features.Products.Commands.CreateProduct;
+using AutoMapper;
+using FluentValidation;
 using MediatR;
 using Microsoft.Extensions.DependencyInjection;
 using System;
@@ -10,11 +12,11 @@ namespace Application
 {
     public static class ServiceRegistration
     {
-        public static IServiceCollection AddApplicationLayer(this IServiceCollection services)
+        public static void AddApplicationLayer(this IServiceCollection services)
         {
             services.AddAutoMapper(Assembly.GetExecutingAssembly());
+            services.AddValidatorsFromAssembly(Assembly.GetExecutingAssembly());
             services.AddMediatR(Assembly.GetExecutingAssembly());
-            return services;
         }
     }
 }