+++ Application/Wrappers/Response.cs
@@ -18,7 +18,7 @@ namespace Application.Wrappers
         }
         public bool Succeeded { get; set; }
 
-        public string[] Errors { get; set; }
+        public List<string> Errors { get; set; }
         public string Message { get; set; }
         public T Data { get; set; }
     }