+++ WebApi/Startup.cs
@@ -1,19 +1,20 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using Application;
+using Application.Features.Products.Commands.CreateProduct;
+using Application.Wrappers;
+using FluentValidation.AspNetCore;
 using Infrastructure.Identity;
 using Infrastructure.Persistence;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.HttpsPolicy;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.ModelBinding;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Logging;
 using Microsoft.OpenApi.Models;
+using System;
+using System.Linq;
+using WebApi.Filters;
 
 namespace WebApi
 {
@@ -31,7 +32,22 @@ namespace WebApi
             services.AddApplicationLayer();
             services.AddIdentityInfrastructure(_config);
             services.AddPersistenceInfrastructure(_config);
-            services.AddControllers();
+            services.AddControllers().AddFluentValidation();
+            services.Configure<ApiBehaviorOptions>(options =>
+            {
+                options.InvalidModelStateResponseFactory = c =>
+                {
+                    var errors = c.ModelState.Values.Where(v => v.Errors.Count > 0)
+                      .SelectMany(v => v.Errors)
+                      .Select(v => v.ErrorMessage).ToList();
+                    return new BadRequestObjectResult(new Response<string>
+                    {
+                        Succeeded = false,
+                        Message = "Validation Errors Occured.",
+                        Errors = errors
+                    });
+                };
+            });
             #region Swagger
             services.AddSwaggerGen(c =>
             {
@@ -80,7 +96,12 @@ namespace WebApi
             {
                 app.UseDeveloperExceptionPage();
             }
-
+            else
+            {
+                app.UseExceptionHandler("/Error");
+                // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
+                app.UseHsts();
+            }
             app.UseHttpsRedirection();
 
             app.UseRouting();